+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Signature of `curry` function.
 - Type signature of `compose` function.
 - Type signature of `alternation` function.
+- Type signature of `tuple` function.
 
 ## [0.11.0] - 2020-12-08